+++ src/imageOperations/multiReceiptsExtractor/multiReceiptsExtractor.ts
@@ -3,7 +3,7 @@ import { MindeeError, MindeeMimeTypeError } from "../../errors";
 import { Polygon } from "../../geometry";
 import { MultiReceiptsDetectorV1 } from "../../product";
 import { ExtractedMultiReceiptImage } from "./extractedMultiReceiptImage";
-import { LocalInputSource } from "../../input/base";
+import { LocalInputSource } from "../../input";
 import { extractFromPage } from "../common";
 import { PositionField } from "../../parsing/standard";
