+++ src/input/index.ts
@@ -1,18 +1,3 @@
 export { PageOptions, PageOptionsOperation } from "./pageOptions";
-export {
-  Base64Input,
-  BytesInput,
-  PathInput,
-  StreamInput,
-  UrlInput,
-  BufferInput,
-} from "./sources";
-export {
-  InputSource,
-  INPUT_TYPE_BASE64,
-  INPUT_TYPE_BYTES,
-  INPUT_TYPE_PATH,
-  INPUT_TYPE_STREAM,
-  INPUT_TYPE_BUFFER,
-} from "./base";
+export * from "./sources";
 export { LocalResponse } from "./localResponse";
