+++ src/input/sources/streamInput.ts
@@ -1,0 +1,35 @@
+import { Readable } from "stream";
+import { LocalInputSource } from "./localInputSource";
+import { INPUT_TYPE_STREAM } from "./inputSource";
+
+interface StreamInputProps {
+  inputStream: Readable;
+  filename: string;
+}
+
+export class StreamInput extends LocalInputSource {
+  private readonly inputStream: Readable;
+  fileObject: Buffer = Buffer.alloc(0);
+
+  constructor({ inputStream, filename }: StreamInputProps) {
+    super({
+      inputType: INPUT_TYPE_STREAM,
+    });
+    this.filename = filename;
+    this.inputStream = inputStream;
+  }
+
+  async init() {
+    this.fileObject = await this.stream2buffer(this.inputStream);
+    this.mimeType = await this.checkMimetype();
+  }
+
+  async stream2buffer(stream: Readable): Promise<Buffer> {
+    return new Promise<Buffer>((resolve, reject) => {
+      const _buf = Array<any>();
+      stream.on("data", (chunk) => _buf.push(chunk));
+      stream.on("end", () => resolve(Buffer.concat(_buf)));
+      stream.on("error", (err) => reject(`Error converting stream - ${err}`));
+    });
+  }
+}
